refactor(engine): split local_copy buffer_pool pool into submodules
